remove await method in favour of awaitOn
Signale codebase hasn't been updated in two years so I created a fork and published it as a package because I love this package but got that sick of VSCode auto importing await every time I wanted to await a promise. I found there was no way to disable this on case by case basis.